backend/requirements.txt
djangorestframework==3.12.2
django-filter==2.4.0
